requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
diffusers==0.34.0
exceptiongroup==1.3.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.18.0
onnxruntime==1.22.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
peft==0.16.0
pillow==11.3.0
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
xformers==0.0.27.post2
zipp==3.23.0
